kmalloc test

+++ src/Rpi lib/testing_lib/helper_macros.h
@@ -2,9 +2,9 @@
 #ifndef __RPI_MACROS_H__
 #define __RPI_MACROS_H__
 
-
+// powers of 2
 #define is_aligned(x, a)        (((x) & ((typeof(x))(a) - 1)) == 0)
-#define is_aligned_ptr(x, a)        is_aligned((uint32_t)x,a)
+#define is_aligned_ptr(x, a)        is_aligned((uint64_t)x,a)
 
 #define pi_roundup(x,n) (((x)+((n)-1))&(~((n)-1)))
